requirements.txt
streamlit
pandas
matplotlib
pdfkit
python-telegram-bot==13.15
jinja2
